GBPlots

Documentation for GBPlots.

GBPlots.plotstatic

GBPlots.plotstatic

GBPlots.plotstatic

GBPlots.plotstatic

GBPlots.plotstatic — Method

plotstatic(type::Type{T}, phenomes::Phenomes; color_scheme::Symbol=:YlGn)::T where {T<:CorHeatPlots}

Trait correlation heatmaps across populations and per population

Examples

phenomes = Phenomes(n=100, t=3); phenomes.entries = string.("entry_", 1:100); phenomes.populations = StatsBase.sample(string.("pop_", 1:5), 100, replace=true); phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(Distributions.MvNormal([1,2,3], LinearAlgebra.I), 100)'; phenomes.phenotypes[1, 1] = missing;
x = plotstatic(CorHeatPlots, phenomes, color_scheme=:viridis)
fnames = saveplots(x)
rm.(fnames)

GBPlots.plotstatic — Method

plotstatic(type::Type{T}, phenomes::Phenomes)::T where {T<:DistributionPlots}

Plot distributions of each trait across populations and per population

Examples

phenomes = Phenomes(n=100, t=3); phenomes.entries = string.("entry_", 1:100); phenomes.populations .= "pop_1"; phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(Distributions.MvNormal([1,2,3], LinearAlgebra.I), 100)'; phenomes.phenotypes[1, 1] = missing;
x = plotstatic(DistributionPlots, phenomes)
fnames = saveplots(x)
rm.(fnames)

GBPlots.plotstatic — Method

plotstatic(type::Type{T}, phenomes::Phenomes; color_scheme::Symbol=:YlGn)::T where {T<:TreePlots}

Plot tree diagrams showing the relationships of each entry using trait information.

Distance metric: Euclidean

Grouping/linkage: Ward's distance

Branch order: Optimal (Bar-Joseph et al, 2001. Bionformatics.)

Examples

phenomes = Phenomes(n=100, t=3); phenomes.entries = string.("entry_", 1:100); phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(Distributions.MvNormal([1,2,3], LinearAlgebra.I), 100)'; 
Ψ::Matrix{Float64} = phenomes.phenotypes;
groups = Clustering.kmeans(Ψ', 5);
phenomes.populations = string.("pop_", groups.assignments); 
phenomes.phenotypes[1, 1] = missing;
x = plotstatic(TreePlots, phenomes)
fnames = saveplots(x)
rm.(fnames)

GBPlots.plotstatic — Method

plotstatic(type::Type{T}, phenomes::Phenomes)::T where {T<:ViolinPlots}

Violin plots across populations and per population across traits

Examples

phenomes = Phenomes(n=100, t=3); phenomes.entries = string.("entry_", 1:100); phenomes.populations = StatsBase.sample(string.("pop_", 1:5), 100, replace=true); phenomes.traits = ["A", "B", "C"]; phenomes.phenotypes = rand(Distributions.MvNormal([1,2,3], LinearAlgebra.I), 100)'; phenomes.phenotypes[1, 1] = missing;
x = plotstatic(ViolinPlots, phenomes)
fnames = saveplots(x)
rm.(fnames)