GBPlots

Documentation for GBPlots.

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot

GBPlots.plot — Method

plot(
    type::Type{T},
    genomes::Genomes;
    n_loci_alleles::Int64 = 1_000,
    seed::Int64 = 42,
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
    rev_label_colors::Bool = false,
    n_threshold_to_show_text::Int64 = 1_000,
)::T where {T<:CorHeatPlots}

Correlation heatmaps:

between loci_alleles across populations

between entries across populations

between loci_alleles per populations

between entries per populations

Note that we are sample n_loci_alleles to use in plotting for computational efficiency. You may use the seed parameter for replicability.

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> hplots = GBPlots.plot(CorHeatPlots, genomes);

julia> fnames = saveplots(hplots, format="png")

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    genomes::Genomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
)::T where {T<:DistributionPlots}

Plot the distribution of allele frequencies from a subset of loci across populations

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> dplots = GBPlots.plot(DistributionPlots, genomes);

julia> fnames = saveplots(dplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    genomes::Genomes;
    n_loci_alleles::Int64 = 1_000,
    seed::Int64 = 42,
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :tol_muted,
)::T where {T<:PCBiPlots}

Using a subset of the allele frequencies, plot the first 2 principal components of the:

entries

loci-alleles

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> pplots = GBPlots.plot(PCBiPlots, genomes);

julia> fnames = saveplots(pplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    genomes::Genomes;
    n_loci_alleles::Int64 = 1_000,
    seed::Int64 = 42,
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :tol_muted,
    horizontal::Bool = true,
)::T where {T<:TreePlots}

Plot tree diagrams showing the relationships of each entry using a subset of the allele frequencies.

Distance metric: Euclidean

Grouping/linkage: Ward's distance

Branch order: Optimal (Bar-Joseph et al, 2001. Bionformatics.)

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> tplots = GBPlots.plot(TreePlots, genomes);

julia> fnames = saveplots(tplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    genomes::Genomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
)::T where {T<:ViolinPlots}

Violin plot of allele frequency distribution (for a subset of loci) per population

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> genomes = extractgenomes(trials); genomes.phenotypes[1,1] = missing;

julia> vplots = GBPlots.plot(ViolinPlots, genomes);

julia> fnames = saveplots(vplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    phenomes::Phenomes;
    plot_size::Tuple{Int64, Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
    rev_label_colors::Bool = false,
)::T where {T<:CorHeatPlots}

Correlation heatmaps:

between traits across populations

between entries across populations

between traits per populations

between entries per populations

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials); phenomes.phenotypes[1,1] = missing;

julia> hplots = GBPlots.plot(CorHeatPlots, phenomes);

julia> fnames = saveplots(hplots, format="png")

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    phenomes::Phenomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
)::T where {T<:DistributionPlots}

Plot distributions of each trait across populations

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials); phenomes.phenotypes[1,1] = missing;

julia> dplots = GBPlots.plot(DistributionPlots, phenomes);

julia> fnames = saveplots(dplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    phenomes::Phenomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :tol_muted,
)::T where {T<:PCBiPlots}

Plot the first 2 principal components of the:

entries

loci-alleles

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> pplots = GBPlots.plot(PCBiPlots, genomes);

julia> fnames = saveplots(pplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    phenomes::Phenomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :tol_light,
    horizontal::Bool = true,
    standardise_traits::Bool = true,
)::T where {T<:TreePlots}

Plot tree diagrams showing the relationships of each entry using trait information.

Distance metric: Euclidean

Grouping/linkage: Ward's distance

Branch order: Optimal (Bar-Joseph et al, 2001. Bionformatics.)

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials); phenomes.phenotypes[1,1] = missing;

julia> tplots = GBPlots.plot(TreePlots, phenomes);

julia> fnames = saveplots(tplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    phenomes::Phenomes;
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
)::T where {T<:ViolinPlots}

Violin plots per trait per population

Examples

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials); phenomes.phenotypes[1,1] = missing;

julia> vplots = GBPlots.plot(ViolinPlots, phenomes);

julia> fnames = saveplots(vplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    cvs::Vector{CV};
    metric::String = "cor",
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
)::T where {T <: BarPlots}

Bar plots summarising the results of a k-fold cross-validation

Examples

julia> cvs::Vector{CV} = [];

julia> for m in 1:3
            for t in 1:5
                for p in 1:4
                    n = 100
                    populations = if rand() < 0.5
                        string.("population_", sample(1:4, n, replace=true))
                    else
                        string.("population_", repeat([p], n))
                    end
                    entries = string.("entry_", sample(1:1_000, n, replace=true))
                    for r in 1:5
                        for f in 1:5
                            fit = Fit(n = 10, l = 1_000); fit.model = string("model_", m); fit.populations .= string("population_", p); fit.trait = string("trait_", t); fit.metrics = Dict("cor" => rand()/maximum([1, 5*rand()]), "rmse" => rand())
                            cv = CV(
                                string("replication_", r), 
                                string("fold_", f), 
                                fit, 
                                populations, 
                                entries, 
                                rand(n), 
                                rand(n), 
                                fit.metrics
                            )
                            push!(cvs, cv)
                        end
                    end
                end
            end
        end;

julia> bplots = GBPlots.plot(BarPlots, cvs);

julia> fnames = saveplots(bplots)

julia> rm.(fnames);

GBPlots.plot — Method

plot(
    type::Type{T},
    cvs::Vector{CV};
    metric::String = "cor",
    plot_size::Tuple{Int64,Int64} = (600, 450),
    colour_scheme::Symbol = :viridis,
)::T where {T <: BoxPlots}

Box plots summarising the results of a k-fold cross-validation

Examples

julia> cvs::Vector{CV} = [];

julia> for m in 1:3
            for t in 1:5
                for p in 1:4
                    n = 100
                    populations = if rand() < 0.5
                        string.("population_", sample(1:4, n, replace=true))
                    else
                        string.("population_", repeat([p], n))
                    end
                    entries = string.("entry_", sample(1:1_000, n, replace=true))
                    for r in 1:5
                        for f in 1:5
                            fit = Fit(n = 10, l = 1_000); fit.model = string("model_", m); fit.populations .= string("population_", p); fit.trait = string("trait_", t); fit.metrics = Dict("cor" => rand()/maximum([1, 5*rand()]), "rmse" => rand())
                            cv = CV(
                                string("replication_", r), 
                                string("fold_", f), 
                                fit, 
                                populations, 
                                entries, 
                                rand(n), 
                                rand(n), 
                                fit.metrics
                            )
                            push!(cvs, cv)
                        end
                    end
                end
            end
        end;

julia> bplots = GBPlots.plot(BoxPlots, cvs);

julia> fnames = saveplots(bplots)

julia> rm.(fnames);